vhd: split a vhd.js file
